refactor(applicant): QueryApplicantService 인터페이스 제거
